Temp+influxdb
I've changed the structure of the influxdb to incorporate temp and bypass status per module.  Temps are now converted to celcius as well.

+++ Arduino-ESP8266-BMS-Controller/bms_values.h
@@ -32,7 +32,8 @@ struct cell_module {
   // 7 bit slave I2C address
   uint8_t address;
   uint16_t voltage;
-  uint16_t temperature;
+//  uint16_t temperature;
+  float temperature;
 
   uint16_t balance_target;
